+++ scripts/pressures_script.py
@@ -1,11 +1,11 @@
 #%%
 # Import Dependencies
 from IPython.display import display_markdown
+from pygeom.geom3d import Vector
 
 from pyapm.classes import panelsystem_from_json
 from pyapm.outputs.msh import panelresult_to_msh
 from pyapm.tools.points import fetch_pids_ttol, point_results
-from pygeom.geom3d import Vector
 
 #%%
 # Create Panel System
@@ -20,7 +20,7 @@ display_markdown(pres)
 
 #%%
 # Determine Point Pressures
-pnts = Vector.zeros((7, 2), dtype=float)
+pnts = Vector.zeros((7, 2))
 pnts[0, 0] = Vector(0.25, 0.05, 0.05)
 pnts[1, 0] = Vector(0.25, 0.05, -0.05)
 pnts[2, 0] = Vector(0.25, 4.5, 0.05)
